Bugfix/ Delivery price not saved correctly to the DB

+++ Services/FinalPoint.Services.Data/Parcel/ParcelService.cs
@@ -47,7 +47,7 @@
                 Weight = input.Weight,
                 NumberOfParts = input.NumberOfParts,
                 HasCashOnDelivery = input.HasCashOnDelivery,
-                CashOnDeliveryPrice = input.CashOnDeliveryPrice,
+                CashOnDeliveryPrice = (double)input.CashOnDeliveryPrice,
                 IsFragile = input.IsFragile,
                 DontPaletize = input.DontPaletize,
                 SendingEmployee = input.SendingEmployee,
